% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenData.R
\name{BetweenData}
\alias{BetweenData}
\title{BetweenData: A function to create time-series cross-section between data}
\usage{
BetweenData(formula, data)
}
\arguments{
\item{formula}{A one sided formula with units on the left hand side; (key ~ .)}

\item{data}{A time series cross-section dataset.}
}
\value{
data.frame A tibble with between data
}
\description{
BetweenData: A function to create time-series cross-section between data
}
\examples{
data(FANG)
BetweenData(symbol ~ ., data=FANG)
}
