% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayWModelFitter.R
\name{DayWModelFitter}
\alias{DayWModelFitter}
\title{Function to estimate daily seasonal time series models}
\usage{
DayWModelFitter(data, Outcome)
}
\arguments{
\item{data}{A tsibble.}

\item{Outcome}{A valid variable name for the Outcome to be modelled in in `data`.}
}
\value{
A mable containing: \itemize{
\item \strong{K=1,2,3}: ARIMA models with fourier(K=1,2,3)
\item \strong{ARIMA, .T, .TS}: an ARIMA model .t (with trend) .ts (with trend and season)
\item \strong{ETS}: an ETS model
\item \strong{NNET}: the model fits of NNETAR with T (trend) TS (trend and season) and numbers indicate K
\item \strong{prophet.LWA}: a prophet model with weekly and annual seasons
\item \strong{prophet.LW}: a prophet model with weekly seasons
\item \strong{Combo1}: the average of ETS and ARIMA
\item \strong{Combo2}: the average of ETS, NNETT2 and ARIMA
}
}
\description{
Function to estimate daily seasonal time series models
}
\examples{
data(SNWSDPDX)
DayWModelFitter(SNWSDPDX, TX)
}
