% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayWModelPicker.R
\name{DayWModelPicker}
\alias{DayWModelPicker}
\title{Function to choose optimal forecast model}
\usage{
DayWModelPicker(data, Outcome, DateVar, H.Horizon = 14)
}
\arguments{
\item{data}{A tsibble.}

\item{Outcome}{A valid variable name for the Outcome to be modelled in in `data`.}

\item{DateVar}{A valid variable name for the time index in `data`.}

\item{H.Horizon}{An integer for the forecast horizon/test subset of `data`.}
}
\value{
A list containing:

Accuracy.Table for the forecast horizon against the test sample.

test: the test set

train: the training set

Model.Fits: the model fits

Model.Forecasts: the forecasts

Min.Model: the minimum model by MAE

Min.Report: the minimum model report

Min.Res.Plot: the gg_tsdisplay for the minimum model

Min.Forecast.Plot: a plot of the minimum forecast by MAE
}
\description{
Function to choose optimal forecast model
}
\examples{
data(SNWSDPDX)
DayWModelPicker(SNWSDPDX, TX, date, H.Horizon=14)
}
