% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FANG}
\alias{FANG}
\title{FANG Stock Data [FANG]}
\format{
An irregular daily tsibble with 4032 rows and 8 variables:
\describe{
  \item{symbol}{a character ticker symbol (1--31)}
  \item{date}{a date [in yy-mm-dd] (2013-01-02--2016-12-30)}
  \item{open}{opening price (22.99--1226.80)}
  \item{high}{high price for the day (23.09--1228.88)}
  \item{low}{low price for the day (22.67--1218.60)}
  \item{close}{closing price for the day (22.9--1220.20)}
  \item{volume}{shares traded for the day (7900--365457900)}
  \item{adjusted}{adjusted closing price for the day (13.14--844.36)}
}
}
\usage{
FANG
}
\description{
An irregular daily tibble (index=date, key=symbol) of stock prices in OHLC for Facebook, Amazon, Netflix, and Google.
}
\keyword{datasets}
