% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb_returns}
\alias{fb_returns}
\title{Facebook Returns [fb_returns]}
\format{
A data frame with 1008 rows and 8 variables:
\describe{
  \item{symbol}{Facebook's ticker ("FB")}
  \item{date}{a calendar date (2013-01-02--2016-12-30)}
  \item{open}{opening price (22.99--133.5)}
  \item{high}{high price for the day (23.09--133.50)}
  \item{low}{low price for the day (0--19)}
  \item{close}{closing price for the day (23.09--133.50)}
  \item{volume}{shares traded for the day (5913100--365457900)}
  \item{adjusted}{adjusted closing price for the day (22.90--133.28)}
  \item{daily.returns}{daily returns for the trading day (-0.069--0.296)}
  \item{trading_day}{the trading day (1--1008)}
}
}
\usage{
fb_returns
}
\description{
A daily dataset of stock price data and returns for Facebook from 2013 through the end of 2016.
}
\keyword{datasets}
