% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonthModelFitter.R
\name{MonthModelFitter}
\alias{MonthModelFitter}
\title{Function to choose optimal forecast model}
\usage{
MonthModelFitter(data, Outcome)
}
\arguments{
\item{data}{A tsibble.}

\item{Outcome}{A valid variable name for the Outcome to be modelled in in `data`.}
}
\value{
A mable containing eleven columns.  3 Fourier and ARIMA models; 1 ARIMA, 1 ETS, 3 NNETAR (fourier 1,2,3) and a prophet
}
\description{
Function to choose optimal forecast model
}
\examples{
data(NWSM)
MonthModelFitter(NWSM, MaxAvg)
}
