% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonthModelPicker.R
\name{MonthModelPicker}
\alias{MonthModelPicker}
\title{Function to choose optimal forecast model for monthly data}
\usage{
MonthModelPicker(data, Outcome, DateVar, H.Horizon = 14)
}
\arguments{
\item{data}{A monthly tsibble.}

\item{Outcome}{A valid variable name for the Outcome to be modelled in in `data`.}

\item{DateVar}{A valid variable name for the time index in `data`.}

\item{H.Horizon}{An integer for the forecast horizon/test subset of `data`.}
}
\value{
A list containing: \itemize{
\item \strong{Accuracy.Table}: accuracy for the forecast horizon against the test sample.
\item \strong{test}: the test set
\item \strong{train}: the training set
\item \strong{Model.Fits}: the model fits
\item \strong{Model.Forecasts}: the forecasts
\item \strong{Min.Model}: the minimum model by MAE
\item \strong{Min.Report}: the minimum model report
\item \strong{Min.Res.Plot}: the gg_tsdisplay for the minimum model
\item \strong{Min.Forecast.Plot}: a plot of the minimum forecast by MAE
}
}
\description{
Function to choose optimal forecast model for monthly data
}
\examples{
data(NWSM)
MonthModelPicker(NWSM, MaxAvg, Month, H.Horizon=12)
}
