% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NWSD}
\alias{NWSD}
\title{Portland, Oregon Daily Weather Data [NWSD]}
\format{
A data frame with 5478 rows and 8 variables:
\describe{
  \item{Day}{an integer Day of the Month (1--31)}
  \item{YR}{four digit calendar year (1941--2019)}
  \item{MO}{two digit calendar month (1--12)}
  \item{TX}{maximum average temperature F (14--107)}
  \item{TN}{maximum average temperature F (-3--74)}
  \item{PR}{Precipitation in in (0--2.69)}
  \item{SN}{Snow in in (0--14.4)}
  \item{date}{a date [Date format] (1941 Jan 1--2019 Dec 31)}
}
}
\usage{
NWSD
}
\description{
A daily tsibble (index=date) for Portland, Oregon containing the daily maximum, minimum temperatures in degrees Fahrenheit and precipitation and snow amounts.
The variables are as follows [as obtained from the national weather service]:
}
\keyword{datasets}
