% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NWSM}
\alias{NWSM}
\title{Portland, Oregon Weather Data [NWSM]}
\format{
A data frame with 240 rows and 7 variables:
\describe{
  \item{MO}{an integer Month (1--12)}
  \item{YR}{four digit calendar year (2000--2019)}
  \item{MaxAvg}{maximum average temperature F (39--87.5)}
  \item{MinAvg}{maximum average temperature F (27.97--61.4)}
  \item{Precip}{Precipitation in in (0--15.24)}
  \item{Snow}{Snow in in (0--19)}
  \item{Month}{a yearmonth time counter (2000 Jan--2019 Dec)}
}
}
\usage{
NWSM
}
\description{
A monthly dataset for Portland, Oregon containing the average monthly maximum, minimum temperatures in degrees Fahrenheit and precipitation and snow amounts.
The variables are as follows [as obtained from the national weather service]:
}
\keyword{datasets}
