% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NWSQ}
\alias{NWSQ}
\title{Portland, Oregon Quarterly Weather Data [NWSQ]}
\format{
A data frame with 316 rows and 5 variables:
\describe{
  \item{Quarter}{yearquarter: a yearquarter (1941 Q1--2019 Q4)}
  \item{TX}{average maximum temperature F (39--87.5)}
  \item{TN}{average maximum temperature F (27.97--61.4)}
  \item{PR}{Total precipitation in in (0--15.24)}
  \item{SN}{Total snow in in (0--19)}
}
}
\usage{
NWSQ
}
\description{
A quarterly dataset for Portland, Oregon containing the daily maximum, minimum temperatures in degrees Fahrenheit and precipitation and snow amounts.
The variables are as follows [as obtained from the national weather service]:
}
\keyword{datasets}
