% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNWSDPDX}
\alias{SNWSDPDX}
\title{Shortened Portland, Oregon Daily Weather Data [SNWSDPDX]}
\format{
A data frame with 1095 rows and 8 variables:
\describe{
  \item{Day}{an integer Day of the Month (1--31)}
  \item{YR}{four digit calendar year (2017--2019)}
  \item{MO}{two digit calendar month (1--12)}
  \item{TX}{maximum average temperature F (39--87.5)}
  \item{TN}{maximum average temperature F (27.97--61.4)}
  \item{PR}{Precipitation in in (0--15.24)}
  \item{SN}{Snow in inches (0--19)}
  \item{date}{a date [Date format] (2017 Jan 1--2019 Dec 31)}
}
}
\usage{
SNWSDPDX
}
\description{
A shortened daily tsibble (index=date) for Portland, Oregon containing the daily maximum, minimum temperatures in degrees Fahrenheit and precipitation and snow amounts for 2017 to 2019.
The variables are as follows [as obtained from the national weather service]:
}
\keyword{datasets}
