% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradingDayCreator.R
\name{TradingDayCreator}
\alias{TradingDayCreator}
\title{Function to create trading days/weeks from a tsibble/grouped tsibble}
\usage{
TradingDayCreator(data, DateVar, key)
}
\arguments{
\item{data}{A tidyquant daily OHLC tibble.}

\item{DateVar}{A time index for the tsibble.}

\item{key}{The key for the tsibble supplied.}
}
\value{
A tsibble indexed by trading_day.  The function will also work with generic messy time units such as weeks.
}
\description{
Function to create trading days/weeks from a tsibble/grouped tsibble
}
\examples{
data(FANG)
TradingDayCreator(FANG, DateVar=date, key=symbol)
}
