% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithinData.R
\name{WithinData}
\alias{WithinData}
\title{WithinData: A function to create time-series cross-section between data}
\usage{
WithinData(formula, data)
}
\arguments{
\item{formula}{A one sided formula with units on the left hand side; (key ~ .)}

\item{data}{A time series cross-section dataset.}
}
\value{
data.frame A tibble with within data
}
\description{
WithinData: A function to create time-series cross-section between data
}
\examples{
data(FANG)
WithinData(symbol ~ ., data=FANG)
}
