% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsum2.R
\name{xtsum2}
\alias{xtsum2}
\title{Time-series cross-section summaries in R
\code{xtsum2} returns three datasets for cross-sectional time series tsibble.}
\usage{
xtsum2(data, key.var)
}
\arguments{
\item{data}{A tibble with clean time and unit(key) indices.}

\item{key.var}{The unit/(key) index.}
}
\value{
a list containing three named elements: Between, Within, and Combined.
}
\description{
Time-series cross-section summaries in R
\code{xtsum2} returns three datasets for cross-sectional time series tsibble.
}
\examples{
data(FANG)
xtsum2(FANG, key.var=symbol)
}
