% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayWModelFitter.R
\name{DayWModelFitter}
\alias{DayWModelFitter}
\title{Function to estimate daily seasonal time series models}
\usage{
DayWModelFitter(data, Outcome)
}
\arguments{
\item{data}{A tsibble.}

\item{Outcome}{A valid variable name for the Outcome to be modelled in in `data`.}
}
\value{
A mable containing: \itemize{
\item \strong{K=1,2,3}: ARIMA models with fourier(K=1,2,3)
\item \strong{ARIMA}: an ARIMA model
\item \strong{ETS}: an ETS model
\item \strong{NNETAR(K=1,2,3)}: the model fits
\item \strong{prophet.LWA}: a prophet model with weekly and annual seasons
\item \strong{prophet.LW}: a prophet model with weekly seasons
\item \strong{Combo1}: the average of ETS and ARIMA
}
}
\description{
Function to estimate daily seasonal time series models
}
\examples{
data(SNWSDPDX)
DayWModelFitter(SNWSDPDX, TX)
}
