% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuarterModelFitter.R
\name{QuarterModelFitter}
\alias{QuarterModelFitter}
\title{Function to estimate quarterly time series models}
\usage{
QuarterModelFitter(data, Outcome)
}
\arguments{
\item{data}{A quarterly tsibble.}

\item{Outcome}{A valid variable name for the Outcome to be modelled in in `data`.}
}
\value{
A mable containing: \itemize{
\item \strong{K=1,2,3}: ARIMA models with fourier(K=1,2,3)
\item \strong{ARIMA}: an ARIMA model
\item \strong{ETS}: an ETS model
\item \strong{NNETAR(K=1,2,3)}: the model fits
\item \strong{prophet}: a prophet model
\item \strong{Combo1}: the average of ETS and ARIMA
}
}
\description{
Function to estimate quarterly time series models
}
\examples{
data(NWSQ)
QuarterModelFitter(NWSQ, TX)
}
